Top level crate w integration tests
