+++ .github/workflows/build.yml
@@ -98,11 +98,11 @@ jobs:
 
       - name: Install dependencies
         run: poetry install
-        working-directory: tests
+        working-directory: tests_py
 
       - name: Initialize the test suite
         run: poetry run poe checks
-        working-directory: tests
+        working-directory: tests_py
 
       - name: Install tools
         run: |
@@ -128,4 +128,4 @@ jobs:
 
       - name: Run test suite
         run: poetry run pytest
-        working-directory: tests
+        working-directory: tests_py
