+++ client/Cargo.toml
@@ -35,8 +35,6 @@ hex = "0.4.3"
 
 
 [dev-dependencies]
-ya-relay-server = { workspace = true }
-
 actix-rt = "2.7"
 actix-web = "4.3"
 bytesize = "1.1"
@@ -62,3 +60,4 @@ env_logger = "0.10.0"
 [features]
 default = []
 packet-trace-enable = ["ya-packet-trace/enable"]
+test-utils = []
