+++ client/src/testing.rs
@@ -1,5 +1,4 @@
 pub mod accessors;
-pub mod init;
 pub mod mocks;
 
 pub use ya_relay_core::session::Session;
