+++ server/Cargo.toml
@@ -46,5 +46,3 @@ serial_test = { git = "https://github.com/tworec/serial_test.git", branch = "act
 ] }
 tokio-stream = "0.1.8"
 test-case = "3.1"
-
-ya-relay-client = { workspace = true }
