+++ server/src/lib.rs
@@ -4,6 +4,5 @@ mod packet;
 mod public_endpoints;
 pub mod server;
 mod state;
-pub mod testing;
 
 pub(crate) use ya_relay_core::error;
