+++ tests/test_discovery.rs
@@ -1,4 +1,4 @@
-pub mod common;
+mod common;
 
 use anyhow::Context;
 use std::rc::Rc;
@@ -9,9 +9,10 @@ use std::time::Duration;
 use ya_relay_client::{ClientBuilder, FailFast};
 use ya_relay_core::crypto::{CryptoProvider, FallbackCryptoProvider};
 use ya_relay_core::key::generate;
-use ya_relay_server::testing::server::init_test_server;
+use common::server::init_test_server;
 
-use common::{hack_make_ip_private, spawn_receive};
+use common::spawn_receive;
+use common::hack_make_ip_private;
 
 #[serial_test::serial]
 async fn test_find_node_by_alias() -> anyhow::Result<()> {
