+++ tests/test_forwarding.rs
@@ -1,4 +1,4 @@
-pub mod common;
+mod common;
 
 use anyhow::Context;
 use futures::StreamExt;
@@ -11,9 +11,10 @@ use tokio_stream::wrappers::UnboundedReceiverStream;
 
 use ya_relay_client::channels::Forwarded;
 use ya_relay_client::{ClientBuilder, FailFast, GenericSender};
-use ya_relay_server::testing::server::init_test_server;
+use common::server::init_test_server;
 
-use common::{hack_make_ip_private, spawn_receive};
+use common::spawn_receive;
+use common::hack_make_ip_private;
 
 #[serial_test::serial]
 async fn test_forward_unreliable() -> anyhow::Result<()> {
